requirements.txt
# Web Framework & API
streamlit>=1.28.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3

# Data Processing & Visualization  
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.17.0
pillow>=10.0.0

# AI Integration (Optional)
google-generativeai>=0.8.0

# Algorithm Dependencies
numpy>=1.24.0
tqdm>=4.65.0
scipy>=1.10.0

# Optimization & Mathematical Libraries
autograd>=1.6.2
cma>=3.2.2

# Development & Testing (Optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.8.0
coverage>=7.3.0

